FIxed bugs in notification page
