+++ PupGo/PupGo.xcodeproj/xcuserdata/yuchenliu.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -35,6 +35,27 @@
 			<key>orderHint</key>
 			<integer>0</integer>
 		</dict>
+		<key>SQLite (Playground) 1.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false/>
+			<key>orderHint</key>
+			<integer>9</integer>
+		</dict>
+		<key>SQLite (Playground) 2.xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false/>
+			<key>orderHint</key>
+			<integer>10</integer>
+		</dict>
+		<key>SQLite (Playground).xcscheme</key>
+		<dict>
+			<key>isShown</key>
+			<false/>
+			<key>orderHint</key>
+			<integer>8</integer>
+		</dict>
 	</dict>
 </dict>
 </plist>
